+++ app/src/main/java/io/agora/flat/data/model/RtcUser.kt
@@ -1,3 +1,3 @@
 package io.agora.flat.data.model
 
-data class RtcUser(val name: String, val rtcUID: Long, val avatarURL: String)
+data class RtcUser(val name: String, val rtcUID: Int, val avatarURL: String)
